feat: enable theme selection
