+++ src/renderer/components/AllRead.test.tsx
@@ -1,5 +1,6 @@
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
+
 import { mockSettings } from '../__mocks__/state-mocks';
 import { ensureStableEmojis } from '../__mocks__/utils';
 import { AppContext } from '../context/App';
