+++ src/renderer/components/Oops.test.tsx
@@ -1,4 +1,5 @@
 import { render } from '@testing-library/react';
+
 import { Oops } from './Oops';
 
 describe('renderer/components/Oops.tsx', () => {
