+++ src/renderer/components/Sidebar.test.tsx
@@ -1,5 +1,6 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
+
 import {
   mockAccountNotifications,
   mockAccountNotificationsWithMorePages,
