+++ src/renderer/components/notifications/NotificationRow.tsx
@@ -58,7 +58,7 @@ export const NotificationRow: FC<INotificationRow> = ({
     <div
       id={notification.id}
       className={cn(
-        'group flex border-b border-gray-100 bg-white px-3 py-2 hover:bg-gray-100',
+        'group flex border-b border-gray-100 px-3 py-2 hover:bg-gray-100 dark:hover:bg-gray-700',
         (isAnimated || animateExit) &&
           'translate-x-full opacity-0 transition duration-[350ms] ease-in-out',
       )}
