+++ src/renderer/components/primitives/Centered.test.tsx
@@ -1,4 +1,5 @@
 import { render } from '@testing-library/react';
+
 import { Centered } from './Centered';
 
 describe('renderer/components/primitives/Centered.tsx', () => {
