+++ src/renderer/components/primitives/EmojiText.test.tsx
@@ -1,4 +1,5 @@
 import { render } from '@testing-library/react';
+
 import { EmojiText, type IEmojiText } from './EmojiText';
 
 describe('renderer/components/primitives/Emoji.tsx', () => {
