+++ src/renderer/components/primitives/Header.test.tsx
@@ -1,4 +1,5 @@
 import { fireEvent, render, screen } from '@testing-library/react';
+
 import { AppContext } from '../../context/App';
 import { Header } from './Header';
