+++ src/renderer/components/settings/AppearanceSettings.tsx
@@ -77,9 +77,9 @@ export const AppearanceSettings: FC = () => {
         </Text>
         <RadioGroup
           options={themeOptions}
-          // defaultValue={settings.theme}
-          defaultValue="LIGHT"
-          isDisabled={true}
+          defaultValue={settings.theme}
+          // defaultValue="LIGHT"
+          // isDisabled={true}
           value={settings.theme}
           onChange={(evt) => {
             updateSetting('theme', evt.target.value as Theme);
