+++ src/renderer/context/App.tsx
@@ -7,6 +7,7 @@ import {
   useMemo,
   useState,
 } from 'react';
+
 import { useInterval } from '../hooks/useInterval';
 import { useNotifications } from '../hooks/useNotifications';
 import {
