+++ src/renderer/hooks/useInterval.ts
@@ -1,7 +1,6 @@
 import { useEffect, useRef } from 'react';
 
 // Thanks to https://overreacted.io/making-setinterval-declarative-with-react-hooks/
-
 export const useInterval = (callback, delay) => {
   const savedCallback = useRef();
