+++ src/renderer/routes/Accounts.test.tsx
@@ -6,6 +6,7 @@ import {
   waitFor,
 } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
+
 import {
   mockAtlassianCloudAccount,
   mockAuth,
