+++ src/renderer/routes/Notifications.tsx
@@ -1,4 +1,5 @@
 import { type FC, useContext, useMemo } from 'react';
+
 import { AllRead } from '../components/AllRead';
 import { Oops } from '../components/Oops';
 import { AccountNotifications } from '../components/notifications/AccountNotifications';
