+++ src/renderer/routes/Settings.test.tsx
@@ -1,5 +1,6 @@
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
+
 import { mockAuth, mockSettings } from '../__mocks__/state-mocks';
 import { AppContext } from '../context/App';
 import { SettingsRoute } from './Settings';
@@ -12,7 +13,6 @@ jest.mock('react-router-dom', () => ({
 
 describe('renderer/routes/Settings.tsx', () => {
   const fetchNotifications = jest.fn();
-  const resetSettings = jest.fn();
 
   afterEach(() => {
     jest.clearAllMocks();
@@ -53,51 +53,5 @@ describe('renderer/routes/Settings.tsx', () => {
 
     expect(fetchNotifications).toHaveBeenCalledTimes(1);
     expect(mockNavigate).toHaveBeenNthCalledWith(1, -1);
-  });
-
-  it('should reset default settings when `confirmed`', async () => {
-    await act(async () => {
-      render(
-        <AppContext.Provider
-          value={{
-            auth: mockAuth,
-            settings: mockSettings,
-            resetSettings,
-          }}
-        >
-          <MemoryRouter>
-            <SettingsRoute />
-          </MemoryRouter>
-        </AppContext.Provider>,
-      );
-    });
-
-    fireEvent.click(screen.getByTestId('settings-reset-defaults'));
-    fireEvent.click(screen.getByTestId('settings-reset-confirm'));
-
-    expect(resetSettings).toHaveBeenCalled();
-  });
-
-  it('should skip reset default settings when `cancelled`', async () => {
-    await act(async () => {
-      render(
-        <AppContext.Provider
-          value={{
-            auth: mockAuth,
-            settings: mockSettings,
-            resetSettings,
-          }}
-        >
-          <MemoryRouter>
-            <SettingsRoute />
-          </MemoryRouter>
-        </AppContext.Provider>,
-      );
-    });
-
-    fireEvent.click(screen.getByTestId('settings-reset-defaults'));
-    fireEvent.click(screen.getByTestId('settings-reset-cancel'));
-
-    expect(resetSettings).not.toHaveBeenCalled();
   });
 });
