+++ src/renderer/utils/api/errors.ts
@@ -1,4 +1,5 @@
 import { AxiosError } from 'axios';
+
 import type { AtlassifyError } from '../../types';
 import { Errors } from '../errors';
 import type { AtlassianAPIError } from './types';
