+++ src/renderer/utils/auth/utils.ts
@@ -1,4 +1,5 @@
 import log from 'electron-log';
+
 import type { Account, AuthState, Token, Username } from '../../types';
 import { getAuthenticatedUser } from '../api/client';
 import { encryptValue } from '../comms';
