+++ src/renderer/utils/comms.ts
@@ -1,4 +1,5 @@
 import { ipcRenderer, shell } from 'electron';
+
 import { defaultSettings } from '../context/App';
 import { type Link, OpenPreference } from '../types';
 import { Constants } from './constants';
