+++ src/renderer/utils/notifications/native.ts
@@ -1,4 +1,5 @@
 import path from 'node:path';
+
 import type {
   AccountNotifications,
   AtlassifyNotification,
